requirements.txt
# Runtime dependencies for ServiceNow MCP server (generated from pyproject.toml)
# Pin strict versions where already pinned; keep others with minimum versions.
# For Azure Web App (zip deploy) include editable install of the local package at end.

mcp[cli]==1.3.0
requests>=2.28.0
pydantic>=2.0.0
python-dotenv>=1.0.0
starlette>=0.27.0
uvicorn[standard]>=0.22.0
httpx>=0.24.0
PyYAML>=6.0

# (Optional) If you need auth code flows, add: authlib>=1.3.0
# (Optional) For production logging improvements, add: structlog>=24.1.0

# Install the local package in editable mode
-e .

# Dev/test tools (uncomment if deploying with them; usually omit in production)
# pytest>=7.0.0
# pytest-cov>=4.0.0
# black>=23.0.0
# isort>=5.12.0
# mypy>=1.0.0
# ruff>=0.0.1
